Local testing tells me there is no test profile
and therefore crashed and makes the cypress
end to end testing fail.

....

Got a error:

/usr/bin/bash: line 162: ./mvnw: Permission denied

I give up on using the wrapper and hopefully it is already installed 🙏
If not I know what the next commit is about.

I do not want to change the permissions on the file given the
container env we are executing the tests...

UsInG ThE WraPeR As PreDiCtEd 🤡

No java 😐

What about this?

This is like debugging my own computer

but it is remote and I have to tear down
and bring up the OS every time :^)

Do we even have Java installed in this image?

ooops