Added docker build for backend to allow e2e testing
The previous versions and attempts to make e2e testing work
tried to make the server run in the same container. This did not work
without some hacky solutions and workarounds. Therefore this
makes a bigger rewrite.

It add's multistage docker builds to make a docker image of our
backend application. This image can we then use as a container
in the Gitlab Services part of the CI. This makes the backend
availble for our Cypress container over the network.

So what is hopefully happening now is:
1. The build stage builds and registers a docker image in the
image repo of our gitlab project.
2. This image is used in the e2e stage for integration testing.
3. Cypress can use the vite server and our java backend to do
front-backend integration testing.

But we never know if this works before we try it.