Switched to different docker builder
Kaniko does not need the elevated privledges.
So it's suggested by other parts of the Gitlab
documentation to use it.

Amended last commit to be more like the docs.

Commented out jobs for faster feedback

Fixed broken syntax

Try less verbose

Try to use only one place for the env