Try to make it work on localhost again
It now works on the CI but was not working
on the localhost? I think.
Anyways, now using the npm run dev
and the spring-boot:run
and the npx cypress run
will now work locally again. But what about
in the CI? That is what this will test!

Maybe not needed?

Try this!