Converged the different URL's to .evn file
With this change we have .env file that is used to build the frontend with the right url.
And also Cypress uses that file to get the right URL to test against.
Everything should still work fine in the nomal way when working on localhost.

Debugging the CI again...

Ooops

more debugg lines.